requirements.txt
Django==2.0.1
django-oauth-toolkit
django-cors-middleware
djangorestframework